refactoring: Separate Logger to Logger and LoggerType
