+++ Source/AppleLogger/PrintLogger.swift
@@ -7,9 +7,9 @@
 
 import Foundation
 
-public final class PrintLogger: Logger {
-    public let logService: PrintLogService
-    public var loggerSetting: LoggerSetting
+public final class PrintLogger: LoggerType {
+    var logService: PrintLogService
+    var loggerSetting: LoggerSetting
     
     public init(_ subsystem: String, category: String) {
         self.loggerSetting = LoggerSetting(subsystem: subsystem, category: category)
@@ -19,5 +19,31 @@ public final class PrintLogger: Logger {
     public init(setting: LoggerSetting) {
         self.loggerSetting = setting
         self.logService = PrintLogService()
+    }
+}
+
+extension PrintLogger: Logger {
+    public var setting: LoggerSetting {
+        return loggerSetting
+    }
+    
+    public func setLogLevel(to level: LogLevel) {
+        loggerSetting.logLevel = level
+    }
+    
+    public func info(_ message: String) {
+        self.log(level: .info, message)
+    }
+    
+    public func debug(_ message: String) {
+        self.log(level: .debug, message)
+    }
+    
+    public func error(_ message: String) {
+        self.log(level: .error, message)
+    }
+    
+    public func fault(_ message: String) {
+        self.log(level: .fault, message)
     }
 }
